# core/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/test": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API测试"
                ],
                "summary": "API测试",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppTestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "测试接口",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/app/account/bind": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "业务应用用户账号存储"
                ],
                "summary": "业务应用用户账号存储",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Bear",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppAccountBindRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务应用用户账号存储",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v1/app/info": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "业务应用列表"
                ],
                "summary": "业务应用列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Bear",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppInfoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务应用用户账号存储",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/response.SingleAppInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/app/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "业务应用登录"
                ],
                "summary": "业务应用登录",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务应用登录",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/app/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "业务应用注册"
                ],
                "summary": "业务应用注册",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务应用注册接口",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v1/app/stat": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "查询业务应用信息"
                ],
                "summary": "查询业务应用信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Bear",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "查询业务应用信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppStatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询业务应用信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/response.SingleAppInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/app/user/ca": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "查询用户证书"
                ],
                "summary": "查询用户证书",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Bear",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "查询用户证书",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppUserCARequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询用户证书",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.SingleCA"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/app/user/register": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "业务应用代理用户注册"
                ],
                "summary": "业务应用代理用户注册",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Bear",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "业务应用代理用户注册",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppUserRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务应用代理用户注册",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.AppAccountBindRequest": {
            "type": "object",
            "properties": {
                "app_counts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.singleCount"
                    }
                },
                "app_name": {
                    "type": "string"
                },
                "did": {
                    "type": "string"
                }
            }
        },
        "request.AppInfoRequest": {
            "type": "object",
            "properties": {
                "app_name": {
                    "type": "string"
                },
                "page_num": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
        "request.AppLoginRequest": {
            "type": "object",
            "properties": {
                "app_account": {
                    "type": "string"
                },
                "app_name": {
                    "type": "string"
                },
                "app_password": {
                    "type": "string"
                }
            }
        },
        "request.AppRegisterRequest": {
            "type": "object",
            "properties": {
                "algo": {
                    "type": "string"
                },
                "app_name": {
                    "type": "string"
                },
                "app_password": {
                    "type": "string"
                },
                "app_phone": {
                    "type": "string"
                },
                "app_type": {
                    "type": "string"
                },
                "app_user": {
                    "type": "string"
                },
                "store_key": {
                    "type": "string"
                }
            }
        },
        "request.AppStatRequest": {
            "type": "object",
            "properties": {
                "app_name": {
                    "type": "string"
                }
            }
        },
        "request.AppTestRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "request.AppUserCARequest": {
            "type": "object",
            "properties": {
                "app_name": {
                    "type": "string"
                },
                "did": {
                    "type": "string"
                }
            }
        },
        "request.AppUserRegisterRequest": {
            "type": "object",
            "properties": {
                "app_name": {
                    "description": "加上AppName 防止篡token与业务应用,在Service层进行校验",
                    "type": "string"
                },
                "nick_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "user_level": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "user_phone": {
                    "type": "string"
                },
                "user_source": {
                    "type": "string"
                },
                "user_source_type": {
                    "type": "string"
                },
                "user_type": {
                    "type": "string"
                }
            }
        },
...
        "request.singleCount": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "response.SingleAppInfo": {
            "type": "object",
            "properties": {
                "app_algo": {
                    "type": "string"
                },
                "app_name": {
                    "type": "string"
                },
                "app_phone": {
                    "type": "string"
                },
                "app_private_md5": {
                    "type": "string"
                },
                "app_public": {
                    "type": "string"
                },
                "app_public_md5": {
                    "type": "string"
                },
                "app_type": {
                    "type": "string"
                },
                "app_user": {
                    "type": "string"
                },
                "store_key": {
                    "type": "string"
                }
            }
        },
        "response.SingleCA": {
            "type": "object",
            "properties": {
                "did": {
                    "type": "string"
                },
                "nick_name": {
                    "type": "string"
                },
                "private_md5": {
                    "description": "Private    string `json:\"private\"`",
                    "type": "string"
                },
                "public": {
                    "type": "string"
                },
                "public_md5": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "user_level": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "user_phone": {
                    "type": "string"
                },
                "user_source": {
                    "type": "string"
                },
                "user_source_type": {
                    "type": "string"
                },
                "user_type": {
                    "type": "string"
                }
            }
        },
...
    }
}